Materials and BlockGroups
Materials will be added to all physical objects in order for it to be
counted as a physical part of the world. Skins will be implemented in
order to customize an object further such as texture. BlockGroups allow
for things to be grouped under one object. The best example for this
would be a fire in a fireplace. If you look at the fireplace it will
say it's a fireplace with a fire in it. Both fireplace and Fire are
ObjectGroups. If you look at the fire in the fireplace it will say a
fire with four burning logs. They work like a tree, but all the objects
remain alone. The fire BlockGroup acts somewhat like an entity, allow
the player to 'take burning log from the fire in the fireplace' or 'get
log from fire'.

I can't wait until I have a working version of this!